tweaked send
send now takes in a data object, to pass along any non visual info that might be wanted. also, some stuff was reorganized. dice now aren't part of the chat-card div because it makes some of the text small, and there's a second chat card div for the footer, otherwise the formatting breaks