CustomRoll Class
Created a custom roll class that takes an item and can conditionally add terms to build the formula, and then returns a better formatted chat message with the results. similar to the previous roll function, but OOP, as the building the formula part was really awkward. This way is cleaner and easier to understand what's happening. Also, handles item formatting rather than having to do that manually, and the promise chaining is better because it actually chains unlike before.