validateFA
validate font awesome icons. Pass in an icon str, and it searches the loaded fa stylesheet and tries to find a fa rule that matches it. That way, we don't have to keep a list of all valid FA icons, and it *should* work if they every update fa to the new version.