CustomDialog class
For building Dialogs. Returns a promise, and that promise will return the result of the callback of whatever button was chosen. So for example, if you have a bunch of sequential dialogs, you can pass the results along down the chain until you get all the results you need. The original dialog only has promises for the simple prompt or confirm, but the custom one didn't return a promise. Also takes in custom buttons, for a smoother experience.