Advantage class
Class with very powerful static methods. An advantage instance is an immutable object that you can check if it has advantage or disadvantage, as well as read the source, and view the icon of the advantage source. It also takes an expired function that  can be run manually, or is run in the Advantage.check function, which takes the "expires" parameter, and determines if the advantage source is still valid or not. Can be as simple as checking x amount of uses, or as complicated as randomness or checking combat state. The passed expires function is given one parameter, which is the advantage instance itself, so you can store info on the advantage object and then reference it in the expires function.

The check function will check utils.tracking.advantage/disadvantage and do a few different things. If there are no sources of disadvantage or advantage, it will return an empty array. If there are sources of disadvantage, and there is an equal amount of advantages, it asks if you want to match to have a neutral roll, and if you have extra, then asks if you want to have +1 advantages active to have advantage on the roll. Lastly, if you have more disadvantages than advantages, it just says "don't bother, it would be a waste", letting you know that you have advantage, but preventing you from wasting them. Finally, returns the array of vantages.

You can then pass that array into pool, and it will create an object for you. That object has a few entries in it, to simplify handling multiple advantages. It will have a formula entry, which is 1d20, 2d20kh, or 2d20kl dependent on if you have advantage or disadvantage. it also returns a 1, 0, or -1 if you want to just check what state you're in, and use that info in some other way. Lastly, it has a label, which is what can be used in the label section of a roll, and it formats the sources and whether it's advantage or not as readably as possible.