rewrite using new classes
check advantage, then take sources, pool them together and get simple data.
play a sound, and once the sound is done, use that data to build a roll, roll it, then take that message, and set the flavor to attack roll and send it to chat. All it needs now is to replace the previous lucky prompt stuff