CustomRoll ordering
tooltips now display in the order of the formula. That means that if the formula is 1d20 + 1, the tooltip will be die, number, rather than the other way around like it was before. Also, iterate checks the prior term and if it's an operator term that was a subtraction term, then the term number is turned negative. That way it shows up as negative in the tooltip.